Moved methods relating to rsvp to another class
