+++ app/src/main/java/com/example/cscb07_project/events/EventDetails.java
@@ -3,16 +3,13 @@ package com.example.cscb07_project.events;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.example.cscb07_project.R;
-import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -27,17 +24,15 @@ public class EventDetails extends AppCompatActivity {
     TextView retrieveEventName, retrieveEventDetails,
             retrieveEventDate, retrieveEventParLim;
     Button retrieveEventRSVP;
-    DatabaseReference mDatabase;
-    FirebaseAuth mAuth;
     String retrieveEventId;
+    RSVPFunctionality rsvpFunctionality;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_event_details);
 
-        mAuth = FirebaseAuth.getInstance();
-        mDatabase = FirebaseDatabase.getInstance().getReference();
+        rsvpFunctionality = new RSVPFunctionality(this);
 
         // Initializes the UI elements
         retrieveEventName = findViewById(R.id.eventDetailName);
@@ -54,7 +49,7 @@ public class EventDetails extends AppCompatActivity {
         retrieveEventRSVP.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                updatingRSVPStatus(retrieveEventId);
+                rsvpFunctionality.updatingRSVPStatus(retrieveEventId);
             }
         });
 
@@ -101,103 +96,4 @@ public class EventDetails extends AppCompatActivity {
             }
         });
     }
-
-    /* In db check to see if there is room available for user to rsvp.
-
-    If yes, set event to true for the user
-            increment the current participants for the event
-            open up EventList class
-            show a toast that it was successful.
-
-    If no, show a toast that there's no room to rsvp.
-    */
-    private void rsvpForEvent(String eventId) {
-        String userId = mAuth.getCurrentUser().getUid();
-        DatabaseReference eventRef = mDatabase.child("Events").child(eventId);
-        eventRef.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                Event event = snapshot.getValue(Event.class);
-                if (event != null) {
-                    int currentParticipants = event.getParticipants();
-                    int participantLimit = event.getParticipantLimit();
-
-                    if (currentParticipants < participantLimit || participantLimit == 0) {
-                        eventRef.child("participants").setValue(currentParticipants + 1);
-                        DatabaseReference userEventRef = mDatabase.child("Users").
-                                child(userId).child("Events").child(eventId);
-                        userEventRef.setValue(true);
-
-                        Intent intent = new Intent(EventDetails.this, EventList.class);
-                        startActivity(intent);
-                        Toast.makeText(EventDetails.this, "RSVP successful", Toast.LENGTH_SHORT).show();
-                    } else {
-                        Toast.makeText(EventDetails.this, "Event is full. Cannot RSVP.", Toast.LENGTH_SHORT).show();
-                    }
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-                Toast.makeText(EventDetails.this, "Error getting event details.", Toast.LENGTH_SHORT).show();
-                throw error.toException();
-            }
-        });
-    }
-
-    /*
-    Does some null checks and checks if it's true that the user has RSVP'ed for the event.
-
-    If yes, remove the event from the users info in db
-            decrement the current participants for the event
-            show a toast that the user has opted out.
-
-    If no, call to function rsvpForEvent
-            (which will handle changing the user's rsvp status to true).
-     */
-    private void updatingRSVPStatus(String eventId) {
-        String userId = mAuth.getCurrentUser().getUid();
-        DatabaseReference userEventRef = mDatabase.child("Users").
-                child(userId).child("Events").child(eventId);
-        userEventRef.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if (snapshot.exists() && snapshot.getValue(Boolean.class) != null && snapshot.getValue(Boolean.class)) {
-                    userEventRef.removeValue();
-                    mDatabase.child("Events").child(eventId).
-                            child("participants").
-                            addListenerForSingleValueEvent(new ValueEventListener() {
-                        @Override
-                        public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                            Long currentParticipants = dataSnapshot.getValue(Long.class);
-                            if (currentParticipants != null && currentParticipants > 0) {
-                                mDatabase.child("Events").child(eventId).
-                                        child("participants").
-                                        setValue(currentParticipants - 1);
-                                Intent intent = new Intent(EventDetails.this, EventList.class);
-                                startActivity(intent);
-                                Toast.makeText(EventDetails.this,
-                                        "You have opted out", Toast.LENGTH_SHORT).show();
-                            }
-                        }
-
-                        @Override
-                        public void onCancelled(@NonNull DatabaseError error) {
-                            Toast.makeText(EventDetails.this, "Error with updating attendance", Toast.LENGTH_SHORT).show();
-                            throw error.toException();
-                        }
-                    });
-                } else {
-                    rsvpForEvent(eventId);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-                Toast.makeText(EventDetails.this, "Error getting RSVP status", Toast.LENGTH_SHORT).show();
-                throw error.toException();
-            }
-        });
-    }
-
 }
